+++ R/res_custom_sk_map.R
@@ -10,7 +10,7 @@
 #' @param shp SpatialPolygons object
 #' @param station_labs logical, should stations be labeled? Defaults to \code{TRUE}
 #' @param lab_loc chr vector of 'R' and 'L', one letter for each station. if no \code{lab_loc} is specified then labels will default to the left.
-#' @param scale_pos scale_pos where should the scale be placed? Options are 'topleft', 'topright', 'bottomleft', or 'bottomright'. Defaults to 'bottomleft'
+###' @param scale_pos scale_pos where should the scale be placed? Options are 'topleft', 'topright', 'bottomleft', or 'bottomright'. Defaults to 'bottomleft'
 #' @param zoom zoom level, 1-21 for stamen maps. Default is to autoscale based on bbox.
 #' @param maptype stamen map type from ggmap::get_stamenmap.  One of c("terrain", "terrain-background", "terrain-labels", "terrain-lines", "toner", "toner-2010", "toner-2011", "toner-background", "toner-hybrid", "toner-labels", "toner-lines", "toner-lite", "watercolor")
 #' #'
@@ -46,19 +46,19 @@
 #' x <- res_custom_sk_map(stations = stns, x_loc = x_coords,
 #'                   sk_result = trnds, y_loc = y_coords,
 #'                   bbox = bounding_elk, lab_loc = lab_dir,
-#'                   scale_pos = pos, shp = shp_fl)
+#'                   shp = shp_fl)
 #'
 #' ### Higher zoom number gives more details, but may not be visible
 #' x_13 <- res_custom_sk_map(stations = stns, x_loc = x_coords,
 #'                   sk_result = trnds, y_loc = y_coords,
 #'                   bbox = bounding_elk, lab_loc = lab_dir,
-#'                   scale_pos = pos, shp = shp_fl, zoom = 13)
+#'                   shp = shp_fl, zoom = 13)
 #'
 #'### Lower zoom number gives coarser text and fewer features
 #' x_11 <- res_custom_sk_map(stations = stns, x_loc = x_coords,
 #'                   sk_result = trnds, y_loc = y_coords,
 #'                   bbox = bounding_elk, lab_loc = lab_dir,
-#'                   scale_pos = pos, shp = shp_fl, zoom = 11)
+#'                    shp = shp_fl, zoom = 11)
 #'
 #'
 #' ### Different maptypes may be used.  All may not be available.
@@ -69,17 +69,17 @@
 #'                   bbox = bounding_elk, lab_loc = lab_dir,
 #'                   scale_pos = pos, shp = shp_fl, maptype = 'terrain')
 #
-y <- res_custom_sk_map <- function(stations
-                              , x_loc
-                              , y_loc
-                              , sk_result = NULL
-                              , bbox
-                              , shp
-                              , station_labs = TRUE
-                              , lab_loc = NULL
-                              , scale_pos = 'bottomleft'
-                              , zoom = NULL
-                              , maptype = 'toner-lite') {
+res_custom_sk_map <- function(stations
+                                   , x_loc
+                                   , y_loc
+                                   , sk_result = NULL
+                                   , bbox
+                                   , shp
+                                   , station_labs = TRUE
+                                   , lab_loc = NULL
+                                   # , scale_pos = 'bottomleft'
+                                   , zoom = NULL
+                                   , maptype = 'toner-lite') {
 
   # check that a shape file exists
   if(class(shp) != 'SpatialPolygons') {
@@ -110,7 +110,6 @@ y <- res_custom_sk_map <- function(stations
   ymax <- max(bbox[c(2,4)])
   bbox <- c(xmin, ymin, xmax, ymax)
 
-
   #check that stations, x_loc, and y_loc match
   if(length(stations) != length(x_loc))
     stop('An incorrect number of x coordinates were specified. One x coordinate must be specified for each station')
@@ -131,162 +130,71 @@ y <- res_custom_sk_map <- function(stations
     # loc$Longitude[loc$Longitude > 0] <- -loc$Longitude[loc$Longitude > 0]
     warning("Positive longitudes given, please double check")
   }
-# convert location info to sf object
-# use lat/lon, WGS84 projection, EPSG:4326.
-loc_sf <- sf::st_as_sf(loc, coords = c("Longitude","Latitude"))
-sf::st_crs(loc_sf) <- 4326
-# Now transform into the projected web-mercator projection EPSG:3857
-#loc_sf <- sf::st_transform(loc_sf, 3857)
+  # convert location info to sf object
+  # use lat/lon, WGS84 projection, EPSG:4326.
+  loc_sf <- sf::st_as_sf(loc, coords = c("Longitude","Latitude"))
+  sf::st_crs(loc_sf) <- 4326
 
-# Define vectors for the colors, shapes and sizes as needed:
-#   1 - 4 are for showing S-K trend results: 1 = increasing, 2 = decreasing,
-#   3 = insignificant, and 4 = insufficient data.
-# This convention holds for colors, shapes and size parameters. The order is
-#   consistent with the original order.
+  # Define vectors for the colors, shapes and sizes as needed:
+  #   1 - 4 are for showing S-K trend results: 1 = increasing, 2 = decreasing,
+  #   3 = insignificant, and 4 = insufficient data.
+  # This convention holds for colors, shapes and size parameters. The order is
+  #   consistent with the original order.
 
-fill_colors <-  c('#444E65', '#A3DFFF', '#247BA0', '#0a0a0a')
-res_point_size <-   c(8,8,8,8)
-res_point_shape <-  c(24, 25, 21, 13)
+  fill_colors <-  c('#444E65', '#A3DFFF', '#247BA0', '#0a0a0a')
+  res_point_size <-   c(8,8,8,8)
+  res_point_shape <-  c(24, 25, 21, 13)
 
-# These are the codes for the fill color, size and shape legends.
-break_vals <- c("inc", "dec", "insig", "insuff")
+  # These are the codes for the fill color, size and shape legends.
+  break_vals <- c("inc", "dec", "insig", "insuff")
 
-# Set background map zoom level automatically if not specified
-if(is.null(zoom)) {
-  diag_size <- sqrt((xmax-xmin)^2 +(ymax-ymin)^2)
-  zoom <- 14 - ceiling(sqrt(10*diag_size))
-  print(paste("Zoom level calculated as", zoom, sep = " "))
-}
-print(paste("maptype is ",maptype))
+  # Set background map zoom level automatically if not specified
+  if(is.null(zoom)) {
+    diag_size <- sqrt((xmax-xmin)^2 +(ymax-ymin)^2)
+    zoom <- 14 - ceiling(sqrt(10*diag_size))
+    print(paste("Zoom level calculated as", zoom, sep = " "))
+  }
+  print(paste("maptype is ",maptype))
 
-bg_map <- ggmap::get_stamenmap(bbox,
-                               maptype = maptype,
-                               source = "stamen",
-                               zoom = zoom,
-                               messaging = FALSE,
-                               epsg = 3785,
-                               urlonly = FALSE)
+  bg_map <- ggmap::get_stamenmap(bbox,
+                                 maptype = maptype,
+                                 source = "stamen",
+                                 zoom = zoom,
+                                 messaging = FALSE,
+                                 epsg = 3785,
+                                 urlonly = FALSE)
 
-m <- ggmap::ggmap(bg_map) +
-  geom_sf(data = shp, aes(), inherit.aes = FALSE,
-          fill = "yellow", col = '#B3B300', alpha = 0.3) +
-  ggthemes::theme_map() +
-  #    geom_sf_text(data = loc_sf, aes(), inherit.aes = FALSE) +
-  geom_sf(data = loc_sf, inherit.aes = FALSE,
-          aes(color = .data$sk_result,
-              fill = .data$sk_result,
-              shape = .data$sk_result,
-              size = .data$sk_result),
-          show.legend = FALSE) +
-  scale_color_manual(values = fill_colors, breaks = break_vals) +
-  scale_fill_manual(values = fill_colors, breaks = break_vals) +
-  scale_size_manual(values = res_point_size, breaks = break_vals) +
-  scale_shape_manual(values = res_point_shape, breaks = break_vals)
+  m <- ggmap::ggmap(bg_map) +
+    geom_sf(data = shp, aes(), inherit.aes = FALSE,
+            fill = "yellow", col = '#B3B300', alpha = 0.3) +
+    ggthemes::theme_map() +
+    #    geom_sf_text(data = loc_sf, aes(), inherit.aes = FALSE) +
+    geom_sf(data = loc_sf, inherit.aes = FALSE,
+            aes(color = .data$sk_result,
+                fill = .data$sk_result,
+                shape = .data$sk_result,
+                size = .data$sk_result),
+            show.legend = FALSE) +
+    scale_color_manual(values = fill_colors, breaks = break_vals) +
+    scale_fill_manual(values = fill_colors, breaks = break_vals) +
+    scale_size_manual(values = res_point_size, breaks = break_vals) +
+    scale_shape_manual(values = res_point_shape, breaks = break_vals)
 
-# # Determine if r and l labs exist
-# if(!is.null(lab_loc)){
-#   if('L' %in% lab_loc){left_labs <- grep('L', lab_loc)}
-#   if('R' %in% lab_loc){right_labs <- grep('R', lab_loc)}
-# } else {
-#   #default to left labels
-#   left_labs <- c(1:length(stations))
-# }
+  if(station_labs) {
+    # Define lat/long for labels, based on stations, alignment, and bbox
+    loc$lab_long <- loc$Longitude + 0.045* loc$align * (bbox[3] - bbox[1])
+    loc$lab_lat <- loc$Latitude + 0.015 * (bbox[4] - bbox[2])
 
-# set map label styles
-# label_style <- list(
-#   "box-shadow" = "none",
-#   "border-radius" = "5px",
-#   "font" = "bold 16px/1.5 'Helvetica Neue', Arial, Helvetica, sans-serif",
-#   "padding" = "1px 5px 1px 5px"
-# )
+    # convert Labels info to sf object, use lat/lon, WGS84 projection, EPSG:4326.
+    labels_sf <- loc %>%
+      select(abbrev, lab_long, lab_lat) %>%
+      sf::st_as_sf(coords = c("lab_long","lab_lat"))
+    sf::st_crs(labels_sf) <- 4326
 
-# # Determine the types of results
-# if('inc' %in% sk_result){inc_icons <- grep('inc', sk_result)}
-# if('dec' %in% sk_result){dec_icons <- grep('dec', sk_result)}
-# if('insig' %in% sk_result){insig_icons <- grep('insig', sk_result)}
+    m <- m +
+      geom_sf_label(data = labels_sf, inherit.aes = FALSE,
+                    aes(label = abbrev))
+  }
 
-# # Plot map
-# m <- leaflet(loc, options = leafletOptions(zoomControl = FALSE), width = 500, height = 500) %>%
-#   addProviderTiles(leaflet::providers$Esri.WorldGrayCanvas) %>%  # Add default OpenStreetMap map tiles, CartoDB.Positron
-#   addPolygons(data = shp, weight = 2, color = '#B3B300', fillColor = 'yellow')
-#
-# if(exists('left_labs')){
-#   m <- m %>%
-#     addLabelOnlyMarkers(lng = ~Longitude[left_labs] * -1, lat = ~Latitude[left_labs]
-#                         , label = loc$abbrev[left_labs]
-#                         , labelOptions = labelOptions(noHide = station_labs
-#                                                       , direction = c('left')
-#                                                       , opacity = 1
-#                                                       , offset = c(-10, 0)
-#                                                       , style = label_style))
-# }
-#
-# if(exists('right_labs')){
-#   m <- m %>%
-#     addLabelOnlyMarkers(lng = ~Longitude[right_labs] * -1, lat = ~Latitude[right_labs]
-#                         , label = loc$abbrev[right_labs]
-#                         , labelOptions = labelOptions(noHide = station_labs
-#                                                       , direction = c('right')
-#                                                       , opacity = 1
-#                                                       , offset = c(10, 0)
-#                                                       , style = label_style))
-# }
-#
-#
-# if(exists('inc_icons')){
-#   # create file path for icon image
-#   ico_loc <- system.file('extdata', 'arrow_inc.png', package = 'SWMPrExtension')
-#
-#   # make icon
-#   icon_img <- makeIcon(iconUrl = ico_loc
-#                        , iconWidth = 30
-#                        , iconHeight = 30
-#                        , iconAnchorX = 15
-#                        , iconAnchorY = 15)
-#
-#   # plot custom icon
-#   m <- m %>%
-#     addMarkers(lng = ~Longitude[inc_icons] * -1, lat = ~Latitude[inc_icons]
-#                , icon = icon_img)
-# }
-#
-# if(exists('dec_icons')){
-#   # create file path for icon image
-#   ico_loc <- system.file('extdata', 'arrow_dec.png', package = 'SWMPrExtension')
-#
-#   # make icon
-#   icon_img <- makeIcon(iconUrl = ico_loc
-#                        , iconWidth = 30
-#                        , iconHeight = 30
-#                        , iconAnchorX = 15
-#                        , iconAnchorY = 15)
-#
-#   # plot custom icon
-#   m <- m %>%
-#     addMarkers(lng = ~Longitude[dec_icons] * -1, lat = ~Latitude[dec_icons]
-#                , icon = icon_img)
-# }
-#
-# if(exists('insig_icons')){
-#   # create file path for icon image
-#   ico_loc <- system.file('extdata', 'bar_insig.png', package = 'SWMPrExtension')
-#
-#   # make icon
-#   icon_img <- makeIcon(iconUrl = ico_loc
-#                        , iconWidth = 40
-#                        , iconHeight = 14
-#                        , iconAnchorX = 20
-#                        , iconAnchorY = 7)
-#
-#   # plot custom icon
-#   m <- m %>%
-#     addMarkers(lng = ~Longitude[insig_icons] * -1, lat = ~Latitude[insig_icons]
-#                , icon = icon_img)
-# }
-#
-# m <- m %>%
-#   addScaleBar(position = scale_pos) %>%
-#   fitBounds(bbox[1], bbox[2], bbox[3], bbox[4])
-
-return(m)
+  return(m)
 }